fixed typos and added error utility
