+++ src/controller/postRouter.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const { createPost } = require('../services/postService');
+const { handleServiceError } = require('../utilities/routerUtilities');
 const { authenticateToken } = require('../middleware/authMiddleware');
 const { validateTextBody, validateScore } = require('../middleware/postMiddleware');
 
@@ -9,29 +10,14 @@ const postRouter = express.Router();
 postRouter.post("/create", authenticateToken, validateTextBody, validateScore, async (req, res) => {
     //TODO check song title exists in API
     try {
-        await createPost(req.user.Username, req.body.Text, req.body.Score, "TTT");
+        await createPost(req.user.Username, req.body.Text, req.body.Score, req.body.Title);
         res.status(201).json({
-            messge: "Post successfully created";
+            messge: "Post successfully created"
         });
     } catch (err) {
         handleServiceError(err, res);
     }
 });
-
-function handleServiceError(error, res) {
-    console.error(error);
-
-    let statusCode = error.name;
-    let message = error.message;
-    if (typeof error.name != "number") {
-        statusCode = 500;
-        message = "Internal Server Error";
-    }
-
-    res.status(statusCode).json({
-        message
-    });
-}
 
 module.exports = {
     postRouter
